backend/requirements.txt
flask
flask-cors
spacy
rapidFuzz
mysql-connector-python
langdetect
requests
smtplib
email
datetime
